added for reputation dimension

+++ front-end/kg-heartbeat/src/utils.js
@@ -17,7 +17,7 @@ function trasform_to_series(quality_data,selectedKGs,quality_dimension,quality_m
             if (series[j].id === quality_data[i].kg_id){
                 const tab_date = quality_data[i].analysis_date.split('-');
                 const date_utc = Date.UTC(parseInt(tab_date[0]),parseInt(tab_date[1])-1,parseInt(tab_date[2]));
-                series[j].data.push([date_utc,parseInt(quality_data[i].Quality_category_array[quality_dimension][quality_metric])])
+                series[j].data.push([date_utc,parseFloat(quality_data[i].Quality_category_array[quality_dimension][quality_metric])])
                 if(series[j].stack === '')
                     series[j].stack = quality_data[i].kg_name;
                 if(series[j].name === '' || series[j].name === undefined)
